refactor: Recreate phone number pattern detector
